Ooops, accidentally broke all other integration tests moving the sleep function somewhere else :-#
